Examples: make both a lib and a bin by default
Note that the trick to use the same file for both was not ideal because
it produces a warning and would compile it twice if it was by default.
So just make both a lib.rs and a main.rs
