+++ examples/printerdemo/rust/Cargo.toml
@@ -14,6 +14,11 @@ license = "MIT"
 path = "main.rs"
 name = "printerdemo"
 
+[lib]
+path = "lib.rs"
+crate-type = ["lib", "cdylib"]
+
+
 [dependencies]
 slint = { path = "../../../api/rs/slint", features = ["backend-android-activity-05"] }
 
@@ -26,14 +31,6 @@ slint-build = { path = "../../../api/rs/build" }
 wasm-bindgen = { version = "0.2" }
 web-sys = { version = "0.3", features=["console"] }
 console_error_panic_hook = "0.1.5"
-
-# Remove the `#wasm#` to uncomment the wasm or android build.
-# This is commented out by default because we don't want to build it as a library by default
-# The CI has a script that does sed "s/#wasm# //" to generate the wasm build.
-
-#wasm# [lib]
-#wasm# path = "main.rs"
-#wasm# crate-type = ["cdylib"]
