+++ examples/printerdemo/rust/index.html
@@ -7,7 +7,6 @@
 <!--
   This is a static html file used to display the wasm build.
   In order to generate the build
-   - uncomment the #wasm# lines in Cargo.toml
    - Run `wasm-pack build --release --target web` in this directory.
 -->
